refactor: use reutilizable components on filters
