+++ src/components/shared/filters/min-episodes.tsx
@@ -7,7 +7,7 @@ export function MinEpisodes() {
   return (
     <div>
       <h5 className="text-md font-semibold text-card-foreground mb-2">{t("library:episode_other")}</h5>
-      <Input type="number" placeholder={t("common:minEpisodes")} min={0} className="bg-muted/50" />
+      <Input type="number" placeholder={t("common:minEpisodes")} min={0} />
     </div>
   );
 }
